% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/AllMethod.R
\docType{methods}
\name{getBIC}
\alias{getBIC}
\alias{getBIC,HMMVB-method}
\alias{getBIC,HMMVBBIC-method}
\title{Accessor for 'BIC' slot.}
\usage{
getBIC(object)

\S4method{getBIC}{HMMVB}(object)

\S4method{getBIC}{HMMVBBIC}(object)
}
\arguments{
\item{object}{Object of class "HMMVB" or "HMMVBBIC".}
}
\description{
This function outputs BIC for a trained HMM-VB model or a vector with BIC values calculated in model selection.
}
