% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getClsid}
\alias{getClsid}
\title{Accessor for 'clsid' slot.}
\usage{
getClsid(object)
}
\arguments{
\item{object}{Object of class "HMMVBclust".}
}
\description{
This function outputs the cluster labels for the object of class HMMVBclust.
}
