% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getClustParam}
\alias{getClustParam}
\title{Accessor for 'clustParam' slot.}
\usage{
getClustParam(object)
}
\arguments{
\item{object}{Object of class "HMMVBclust".}
}
\description{
This function outputs clusterPar for the object of class HMMVBclust.
}
