% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getConfigList}
\alias{getConfigList}
\title{Accessor for 'configList' slot.}
\usage{
getConfigList(object)
}
\arguments{
\item{object}{Object of class "HMMVBBIC".}
}
\description{
This function outputs list with configurations for the number of states in variable blocks used in model selection.
}
