% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getDiagCov}
\alias{getDiagCov}
\title{Accessor for 'diagCov' slot.}
\usage{
getDiagCov(object)
}
\arguments{
\item{object}{Object of class "HMMVB".}
}
\description{
This function outputs diagCov logical indicator of diagonal covariance 
matrices for HMM-VB model.
}
