% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getHmmChain}
\alias{getHmmChain}
\title{Accessor for 'HmmChain' slot.}
\usage{
getHmmChain(object)
}
\arguments{
\item{object}{Object of class "HMMVB".}
}
\description{
This function outputs a list with trained HMMs.
}
