% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getHmmParam}
\alias{getHmmParam}
\title{Accessor for parameters of HMM}
\usage{
getHmmParam(object)
}
\arguments{
\item{object}{Object of class "HMM".}
}
\description{
This function outputs a list with means, covariance matrices, inverse covarince matrices and
logarithms of the determinants of the covariance matrices for all states of the HMM.
}
