% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getOptHMMVB}
\alias{getOptHMMVB}
\title{Accessor for 'optHMMVB' slot.}
\usage{
getOptHMMVB(object)
}
\arguments{
\item{object}{Object of class "HMMVBBIC".}
}
\description{
This function outputs the optimal HMM-VB found via BIC model selection.
}
