% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getPrenumst}
\alias{getPrenumst}
\title{Accessor for 'prenumst' slot}
\usage{
getPrenumst(object)
}
\arguments{
\item{object}{Object of class "HMM".}
}
\description{
This function outputs the number of states in the HMM for the preceding block of HMM-VB.
}
