% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getSize}
\alias{getSize}
\title{Accessor for 'size' slot.}
\usage{
getSize(object)
}
\arguments{
\item{object}{Object of class "HMMVBclust".}
}
\description{
This function outputs the number of points in each cluster for the object of class HMMVBclust.
}
