% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getVb}
\alias{getVb}
\title{Accessor for 'VbStructure' slot.}
\usage{
getVb(object)
}
\arguments{
\item{object}{Object of class "HMMVB".}
}
\description{
This function outputs the variable block structure in the HMM-VB.
}
