% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{hmm}
\alias{hmm}
\title{Make an instance of "HMM" class.}
\usage{
hmm(dim, numst, prenumst, a00, a, mean, sigma, sigmaInv, sigmaDetLog)
}
\arguments{
\item{dim}{Dimensionality of the data in HMM.}

\item{numst}{An integer vector specifying the number of HMM states.}

\item{prenumst}{An integer vector specifying the number of states of previous 
variable block HMM.}

\item{a00}{Probabilities of HMM states.}

\item{a}{Transition probability matrix from states in the previous variable block
to the states in the current one.}

\item{mean}{A numerical matrix with state means. \emph{k}th row corresponds to the
\emph{k}th state.}

\item{sigma}{A list containing state covariance matrices.}

\item{sigmaInv}{A list containing state inverse covariance matrices.}

\item{sigmaDetLog}{A vector with \eqn{log(|sigma|)} for each state.}
}
\value{
An object of class 'HMM'.
}
\description{
This function creates a Hidden Markov Model on a variable block.
}
\keyword{internal}
