% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{mkhmmvb}
\alias{mkhmmvb}
\title{Make an instance of "HMMVB" class.}
\usage{
mkhmmvb(VbStructure, HmmChain, BIC, diagCov, Loglikehd)
}
\arguments{
\item{VbStructure}{An object of class 'VB' that contains variable block structure.}

\item{HmmChain}{A list of objects of class 'HMM' with trained Hidden Markov Models
for each variable block.}

\item{BIC}{BIC value for provided variable block structure or optimal BIC value
for found variable block structure.}

\item{diagCov}{A logical value indicating whether or not covariance matrices 
for mixture models are diagonal.}

\item{Loglikehd}{Loglikelihood value for each data point}
}
\value{
An instance of class "HMMVB"
}
\description{
This function creates an instance of "HMMVB" class. The function is called inside 
C code during HMMVB model training. It is not meant to be called by users.
}
\keyword{internal}
