% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\docType{data}
\name{sim2}
\alias{sim2}
\title{Synthetic dataset used in section 5.1.2 of the reference paper.}
\format{A data frame with 5000 rows and 5 variables. Last column contains ground truth cluster labels.}
\usage{
sim2
}
\description{
Dataset used for testing clustering with HMM-VB.
The data dimension is 5. Data points were drawn from a 10-component Gaussian Mixture Model.
By specific choice of the means, the data contains 10 distinct clusters. For details see the 
references.
}
\references{
Lin Lin and Jia Li, "Clustering with hidden Markov model on variable blocks," \strong{Journal of Machine Learning Research}, 18(110):1-49, 2017.
}
\keyword{datasets}
